chore: New endpoints, billing rules update. Edit instead of ADD+DELETE.
